Rename api/server to server
